big PR changes
https://coveord.atlassian.net/browse/JSUI-2471
